Fixed styles issue
